requirements.txt
streamlit==1.30.0
openai==2.00.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.4
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
huggingface_hub==0.19.4
transformers==4.37.0
protobuf==4.23.4
tqdm==4.66.2
scipy==1.11.4
requests==2.31.0
